updates example chainlet
