+++ docs/chains/getting-started.mdx
@@ -37,7 +37,7 @@ More details are in the [concepts section](/chains/concepts).
 Create a Chain project directory with a python file in it. You can chose a name
 and location, in this example we assume the file is named `hello.py`.
 
--- Note: If you are changing this snippet, please update the starter code in cli.py accordingly
+-- Note: If you are changing this snippet, please update the example code in example_chainlet.py accordingly
 ```python
 import random
 import truss_chains as chains
