+++ truss/cli/cli.py
@@ -64,33 +64,6 @@ click.rich_click.COMMAND_GROUPS = {
     ]
 }
 
-# Note: If you are changing this snippet, please update the example at docs/chains/getting_started.mdx
-HELLO_PY = """import random
-# for more on chains, check out https://truss.baseten.co/chains/intro
-import truss_chains as chains
-
-
-# By inhereting chains.ChainletBase, chains will know to create a chainlet that hosts the RandInt class
-class RandInt(chains.ChainletBase):
-
-    # run_remote must be implemented by all chainlets. This is the code that will be executed at inference time
-    def run_remote(self, max_value: int) -> int:
-        return random.randint(1, max_value)
-
-# The @chains.mark_entrypoint decorator indicates that this Chainlet is the entrypoint.
-# There must be exactly one entrypoint per chainlet
-@chains.mark_entrypoint
-class HelloWorld(chains.ChainletBase):
-    # chains.depends indicates that the HelloWorld chainlet depends on the RandInt Chainlet
-    # this enables the HelloWorld chainlet to call the RandInt chainlet
-    def __init__(self, rand_int=chains.depends(RandInt, retries=3)) -> None:
-        self._rand_int = rand_int
-
-    def run_remote(self, max_value: int) -> str:
-        num_repetitions = self._rand_int.run_remote(max_value)
-        return "Hello World! " * num_repetitions
-"""
-
 
 def echo_output(f: Callable[..., object]):
     @wraps(f)
@@ -471,15 +444,31 @@ def chains_init(
     user_friendly_path = os.path.join(target_directory, FILENAME)
     rich.print(f"Creating {user_friendly_path}...\n")
 
-    os.mkdir(abs_path)
+    source_code = _load_example_chainlet_code()
+
+    if not os.path.exists(abs_path):
+        os.mkdir(abs_path)
     with open(filename, "w") as f:
-        f.write(HELLO_PY)
+        f.write(source_code)
 
     rich.print(
         "Next steps:\n",
-        f"Run `python {user_friendly_path}` locally by adding `chains.run_local()` to your main.py file\n",
+        f"Run `python {user_friendly_path}` to execute the code locally\n",
         f"Run `truss chains deploy {user_friendly_path}` to deploy the chain to Baseten\n",
     )
+
+
+def _load_example_chainlet_code() -> Optional[str]:
+    try:
+        from truss_chains import example_chainlet
+    # If `example_chainlet` fails validation and `_load_example_chainlet_code` is
+    # called as a result of that, we have a circular import ("partially initialized
+    # module 'truss_chains.example_chainlet' ...").
+    except Exception as e:
+        raise Exception("failed to load starter code. Please notify support") from e
+
+    source = Path(example_chainlet.__file__).read_text()
+    return source
 
 
 def _extract_and_validate_model_identifier(
